# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Seedling Desktop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /screening/plates:
    get:
      operationId: listScreeningPlates
      tags: [screening]
      summary: List screening data by plate
      parameters:
        - name: berryId
          in: query
          schema:
            type: integer
        - name: teamId
          in: query
          schema:
            type: integer
        - name: pollinationYear
          in: query
          schema:
            type: integer
        - name: progeny
          in: query
          schema:
            type: string
        - name: program
          in: query
          schema:
            type: string
        - name: testingLab
          in: query
          schema:
            type: string
        - name: plateIndex
          in: query
          schema:
            type: integer
        - name: screening
          in: query
          schema:
            type: boolean
        - name: sorted
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of screening plates
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScreeningPlate"
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  totalPages:
                    type: integer
  /screening/plates/totals:
    get:
      operationId: getScreeningPlateTotals
      tags: [screening]
      summary: Get screening plate totals
      parameters:
        - name: berryId
          in: query
          schema:
            type: integer
        - name: teamId
          in: query
          schema:
            type: integer
        - name: pollinationYear
          in: query
          schema:
            type: integer
        - name: progeny
          in: query
          schema:
            type: string
        - name: program
          in: query
          schema:
            type: string
        - name: testingLab
          in: query
          schema:
            type: string
        - name: plateIndex
          in: query
          schema:
            type: integer
        - name: screening
          in: query
          schema:
            type: boolean
        - name: sorted
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Screening plate totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreeningPlateTotals"
  /screening/progeny:
    get:
      operationId: listScreeningProgeny
      tags: [screening]
      summary: List screening data by progeny
      parameters:
        - name: berryId
          in: query
          schema:
            type: integer
        - name: teamId
          in: query
          schema:
            type: integer
        - name: pollinationYear
          in: query
          schema:
            type: integer
        - name: progeny
          in: query
          schema:
            type: string
        - name: program
          in: query
          schema:
            type: string
        - name: spCrosses
          in: query
          schema:
            type: boolean
        - name: screening
          in: query
          schema:
            type: boolean
        - name: sorted
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of screening progeny
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScreeningProgeny"
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  totalPages:
                    type: integer
  /screening/progeny/totals:
    get:
      operationId: getScreeningProgenyTotals
      tags: [screening]
      summary: Get screening progeny totals
      parameters:
        - name: berryId
          in: query
          schema:
            type: integer
        - name: teamId
          in: query
          schema:
            type: integer
        - name: pollinationYear
          in: query
          schema:
            type: integer
        - name: progeny
          in: query
          schema:
            type: string
        - name: program
          in: query
          schema:
            type: string
        - name: spCrosses
          in: query
          schema:
            type: boolean
        - name: screening
          in: query
          schema:
            type: boolean
        - name: sorted
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Screening progeny totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreeningProgenyTotals"
...
components:
  schemas:
...
    ScreeningPlate:
      type: object
      properties:
        id:
          type: integer
        plateIndex:
          type: integer
        progeny:
          type: string
        testingLab:
          type: string
          nullable: true
        samplesRequired:
          type: integer
          nullable: true
        samplesCollected:
          type: integer
          nullable: true
        sampleCollectionDate:
          type: string
          nullable: true
        totalKeepRequest:
          type: integer
          nullable: true
        totalKeepActual:
          type: integer
          nullable: true
        totalDiscardsActual:
          type: integer
          nullable: true
        discardDate:
          type: string
          nullable: true
        sorted:
          type: boolean
        sortGroup1:
          type: integer
          nullable: true
        sortGroup2:
          type: integer
          nullable: true
        sortGroup3:
          type: integer
          nullable: true
        sortGroup4:
          type: integer
          nullable: true
        sortGroup5:
          type: integer
          nullable: true
        screening:
          type: boolean
        berry:
          type: string
          nullable: true
        berryId:
          type: integer
          nullable: true
        teamName:
          type: string
          nullable: true
        teamId:
          type: integer
          nullable: true
        d1Program:
          type: string
          nullable: true
        pollinationYear:
          type: integer
          nullable: true
      required:
        - id
        - progeny
    ScreeningPlateTotals:
      type: object
      properties:
        rowCount:
          type: integer
        samplesRequired:
          type: number
        samplesCollected:
          type: number
        totalKeepRequest:
          type: number
        totalKeepActual:
          type: number
        totalDiscardsActual:
          type: number
        sortGroup1:
          type: number
        sortGroup2:
          type: number
        sortGroup3:
          type: number
        sortGroup4:
          type: number
        sortGroup5:
          type: number
    ScreeningProgeny:
      type: object
      properties:
        id:
          type: integer
        progeny:
          type: string
        d1Program:
          type: string
          nullable: true
        d2Program:
          type: string
          nullable: true
        totalPlatesRequired:
          type: integer
          nullable: true
        platesCollected:
          type: integer
          nullable: true
        sampleRequired:
          type: integer
          nullable: true
        samplesCollected:
          type: integer
          nullable: true
        keepRequest:
          type: integer
          nullable: true
        keepActual:
          type: integer
          nullable: true
        totalDiscardsActual:
          type: integer
          nullable: true
        sorted:
          type: boolean
        sortGroup1:
          type: integer
          nullable: true
        sortGroup2:
          type: integer
          nullable: true
        sortGroup3:
          type: integer
          nullable: true
        sortGroup4:
          type: integer
          nullable: true
        sortGroup5:
          type: integer
          nullable: true
        testingLab1:
          type: string
          nullable: true
        testingLab2:
          type: string
          nullable: true
        screening:
          type: boolean
        berry:
          type: string
          nullable: true
        berryId:
          type: integer
          nullable: true
        teamName:
          type: string
          nullable: true
        teamId:
          type: integer
          nullable: true
        pollinationYear:
          type: integer
          nullable: true
      required:
        - id
        - progeny
    ScreeningProgenyTotals:
      type: object
      properties:
        rowCount:
          type: integer
        sampleRequired:
          type: number
        samplesCollected:
          type: number
        keepRequest:
          type: number
        keepActual:
          type: number
        totalDiscardsActual:
          type: number
        sortGroup1:
          type: number
        sortGroup2:
          type: number
        sortGroup3:
          type: number
        sortGroup4:
          type: number
        sortGroup5:
          type: number
...